python: Set Py_mod_gil in rti1516 module initialization.
According to PEP 703 - Making the Global Interpreter Lock
Optional in CPython, the python interface should not be prone to
thread safety issues. The module does not use container
access methods that return borrowed references as 'Borrowed References'
section in the above PEP. When calling back into python, new containers
are assembled from C++ and when finished provided as arguments to
python code. This should be safe as assembling is finished before
calling into python.
Nevertheless, the wrapped rti library is not considered threadsafe in
itself. The way the rti api is meant to be used follows the same threading
guarantees that the backing c++ library provides. That is, accessing the
same RTIambassador instance is not thread safe or atomic in itself and
concurrent access to a single RTIambassador instance needs to be
serialized in c++ as well as it needs to be in python.