python: Use heap type objects in the rti1516 module.
It turns out that the recent gil work with python
just requires the use of heap type objects, even
for the use case that we have. The data that is
shared among multiple interpreters are the method
descriptors. The way the static type objects in
the python core library uses special code to
store a dict per interpreter. But that code is
not accessible from outside the core code.
Finally we need to use heap type objects,
even if the change is intrusive. But the ability to
have several interpreters running concurrently
outweights the intrusive change.